pyproject.toml
[project]
name = "network-shaping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "gymnasium>=0.29",
  "stable-baselines3>=2.0",
  "numpy>=1.24",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
  "seaborn>=0.12",
]
